backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# AI/ML Libraries
openai==1.10.0
anthropic>=0.40.0
stability-sdk==0.8.4
elevenlabs==0.2.27
replicate==0.20.0

# Media Processing
opencv-python==4.9.0.80
moviepy==1.0.3
Pillow==10.2.0
numpy==1.26.3

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Storage
boto3==1.34.34
azure-storage-blob==12.19.0

# Task Queue
celery==5.3.6
redis==5.0.1

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.13.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
